refactor: majority of EntryEditor fields now operate via react-hook-form
