+++ package.json
@@ -14,6 +14,7 @@
     "react-bootstrap": "^1.4.3",
     "react-datetime-picker": "^3.0.4",
     "react-dom": "^18.1.0",
+    "react-hook-form": "^7.38.0",
     "react-icons": "^3.10.0",
     "react-moment": "^1.1.2",
     "react-redux": "^8.0.2",
