+++ src/components/Attachment/Attachment.js
@@ -20,28 +20,25 @@ import Image from 'react-bootstrap/Image';
 import Button from 'react-bootstrap/Button';
 import { FaRegFile } from "react-icons/fa";
 
-class Attachment extends Component{
+const Attachment = ({attachment, removeAttachment}) => {
 
-    render(){
-        if(this.props.file.file.type.toLowerCase().startsWith("image")){
-            return(
-                <div className="attachment">
-                    <Button variant="danger" onClick={() => this.props.removeAttachment(this.props.file)}>Remove</Button>
-                    <Image src={URL.createObjectURL(this.props.file.file)} className="attachment"/>
-                    <p>{this.props.file.file.name}</p>
-                </div>
-            )
-        }
-        else{
-            return(
-                <div className="attachment">
-                    <Button variant="danger" onClick={() => this.props.removeAttachment(this.props.file)}>Remove</Button><br/>
-                    <FaRegFile style={{marginTop: "5px"}} size={56}/>
-                    <p>{this.props.file.file.name}</p>
-                </div>
-            )
-        }
-        
+    if(attachment.file.type.toLowerCase().startsWith("image")){
+        return(
+            <div className="attachment">
+                <Button variant="danger" onClick={() => removeAttachment(attachment.file)}>Remove</Button>
+                <Image src={URL.createObjectURL(attachment.file)} className="attachment"/>
+                <p>{attachment.file.name}</p>
+            </div>
+        )
+    }
+    else{
+        return(
+            <div className="attachment">
+                <Button variant="danger" onClick={() => removeAttachment(attachment.file)}>Remove</Button><br/>
+                <FaRegFile style={{marginTop: "5px"}} size={56}/>
+                <p>{attachment.file.name}</p>
+            </div>
+        )
     }
 }
